making validations and extracts orderable. Optimizing and adding information to the settings menu
